[jsonwriter] fix 32/64 ambiguity  [metadata/Tables] Fix bad enum usage
Former-commit-id: 69042e44684e70e85a2b50a68763ad31e6ef8370 [formerly 97365620041a8938becbbc0ada58952ab72406fb] [formerly 69042e44684e70e85a2b50a68763ad31e6ef8370 [formerly 97365620041a8938becbbc0ada58952ab72406fb] [formerly 97365620041a8938becbbc0ada58952ab72406fb [formerly bd049992e5122091f41c62367e649011c338463a]]]
Former-commit-id: 588ca9b6729861060762f189043d09cdd4b76078
Former-commit-id: 8988ea3b1abe0d09020d8954926fe9efe0886f56
Former-commit-id: eca90d3fa121f69173cd1768fe273d0f07d33ad6

+++ src/metadata_enrichment/Tables.cpp
@@ -1,5 +1,6 @@
 #include "metadata_enrichment/Tables.h"
 
+#include <cstdint>
 #include <dbproxy/dbproxy.h>
 
 #include <json/jsonWriter.h>
@@ -220,7 +221,7 @@ namespace wasabi{
         return Column::Aggregation::None;
       }
       stringstream aStream;
-      aStream << "datatype " << theType << " not found" << endl;
+      aStream << "datatype " << static_cast<uint8_t>(theType) << " not found" << endl;
       throw out_of_range(aStream.str());
     }
     Column::DataType calcDataType( const string& theType)
@@ -274,8 +275,8 @@ namespace wasabi{
       JSON_MAP(theWriter);
       theWriter.pair("Name", getName());
       theWriter.pair("SQLName", getSQLName());
-      theWriter.pair("Aggregation", getAggregation());
-      theWriter.pair("DataType",getDataType());
+      theWriter.pair("Aggregation", static_cast<std::uint32_t>(getAggregation()));
+      theWriter.pair("DataType",static_cast<std::uint32_t>(getDataType()));
     }
     ostream& operator<<(ostream& theStream, const Column& theColumn)
 	{
